refactor: 연관된 엔티티끼리 패키지화
