+++ src/main/java/org/omo/omospringboot/entity/location/TouristSpot.java
@@ -1,4 +1,4 @@
-package org.omo.omospringboot.entity;
+package org.omo.omospringboot.entity.location;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.DiscriminatorValue;
@@ -6,17 +6,19 @@ import jakarta.persistence.Entity;
 import jakarta.persistence.Table;
 import lombok.*;
 
+import java.time.LocalDateTime;
+
 @Entity
 @Table
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor
 @Builder
-@DiscriminatorValue("LEISURE")
-public class Leisure extends Location{
-    private String duration; // 개장기간
+@DiscriminatorValue("TOURIST_SPOT")
+public class TouristSpot extends Location{
+    private LocalDateTime startDay;
 
-    private String fee;
+    private String experienceInformation;
 
     private String pet;
