+++ src/main/java/org/omo/omospringboot/entity/location/UniqueExperience.java
@@ -1,6 +1,5 @@
-package org.omo.omospringboot.entity;
+package org.omo.omospringboot.entity.location;
 
-import jakarta.persistence.Column;
 import jakarta.persistence.DiscriminatorValue;
 import jakarta.persistence.Entity;
 import jakarta.persistence.Table;
