+++ src/main/java/org/omo/omospringboot/provider/JwtTokenProvider.java
@@ -7,9 +7,9 @@ import io.jsonwebtoken.security.SecurityException;
 import lombok.extern.slf4j.Slf4j;
 import org.omo.omospringboot.constant.ErrorCode;
 import org.omo.omospringboot.constant.TokenType;
-import org.omo.omospringboot.entity.User;
+import org.omo.omospringboot.entity.user.User;
 import org.omo.omospringboot.exception.CustomErrorException;
-import org.omo.omospringboot.repository.UserRepository;
+import org.omo.omospringboot.repository.user.UserRepository;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
